// src/API.ts
export enum TodoStatus {
  NONE = "NONE",
  BACKLOG = "BACKLOG",
  DOING = "DOING",
  REVIEW = "REVIEW",
  DONE = "DONE",
}
// ...
export type GetTodoQuery = {
  getTodo:  {
    __typename: "Todo",
    id: string,
    name: string,
    status: TodoStatus,
    deadLine: string | null,
    description: string | null,
    createdAt: string,
    updatedAt: string,
    owner: string | null,
  } | null,
};